+++ modules/gov/utils_proposal.go
@@ -3,6 +3,7 @@ package gov
 import (
 	"fmt"
 	"strings"
+	"time"
 
 	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
 	proposaltypes "github.com/cosmos/cosmos-sdk/x/params/types/proposal"
@@ -20,11 +21,14 @@ import (
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
-func (m *Module) UpdateProposal(height int64, id uint64) error {
+func (m *Module) UpdateProposal(height int64, blockTime time.Time, id uint64) error {
 	// Get the proposal
 	proposal, err := m.source.Proposal(height, id)
 	if err != nil {
-		if strings.Contains(err.Error(), codes.NotFound.String()) {
+		// Check if proposal has reached the voting end time
+		passedVotingPeriod := blockTime.After(proposal.VotingEndTime)
+
+		if strings.Contains(err.Error(), codes.NotFound.String()) && passedVotingPeriod {
 			// Handle case when a proposal is deleted from the chain (did not pass deposit period)
 			return m.updateDeletedProposalStatus(id)
 		}